+++ src/PSCompression/Extensions/MiscExtensions.cs
@@ -1,8 +1,10 @@
 using System.Collections.Generic;
+using System.Net;
+using System.Security;
 
 namespace PSCompression.Extensions;
 
-internal static class DictionaryExtensions
+internal static class MiscExtensions
 {
     internal static void Deconstruct<TKey, TValue>(
         this KeyValuePair<TKey, TValue> keyv,
@@ -12,4 +14,7 @@ internal static class DictionaryExtensions
         key = keyv.Key;
         value = keyv.Value;
     }
+
+    internal static string AsPlainText(this SecureString secureString) =>
+        new NetworkCredential(string.Empty, secureString).Password;
 }